fix: style 정비
