+++ app/app.css
@@ -10,6 +10,7 @@
   --color-light-blue: #5988ff;
   --color-bg-0: #121212;
   --color-white-87: rgba(255, 255, 255, 0.87);
+  --color-white-15: rgba(255, 255, 255, 0.15);
   --leading-normal: 1.2;
 }
 
@@ -54,11 +55,7 @@ body {
   --color-sidebar-accent-foreground: var(--sidebar-accent-foreground);
   --color-sidebar-border: var(--sidebar-border);
   --color-sidebar-ring: var(--sidebar-ring);
-  --gradient-bg-5: linear-gradient(
-    0deg, 
-    oklch(1 0 0 / 0.05) 0%, 
-    oklch(1 0 0 / 0.05) 100%
-  ), oklch(0.11 0 0);
+  --gradient-bg-5: linear-gradient(0deg, oklch(1 0 0 / 0.05) 0%, oklch(1 0 0 / 0.05) 100%), oklch(0.11 0 0);
 }
 
 :root {
